feat(cobordism): remove stage-2 length clamping entirely — unbounded Lorentzian trials
Addresses the PR review: no clamping at all, anywhere — not the causal-aware
guard band, not the baseline [0.05, 20] spacelike clamp it wrapped. Every
stage-2 trial l^2 is written unbounded: spacelike, timelike, and lightlike
(any (-eps, eps) band) trials are all admissible, on both cone sides, at any
magnitude. The only rejection is the line search's own: a trial the objective
cannot evaluate scores +inf and is backed off, so degenerate geometry is
refused by the physics, never by a projection. Deletes setCausalGuard /
causalGuardEpsilon / boundedTrialRealPart / kDegeneracyFloor / kMagnitudeCap
and their bindings; not backwards compatible by design.

The guard test suite becomes test_stage2_unclamped_python.py: the projection
API must be GONE, accepted steps never pin an edge to the retired clamp
values, the old floor band is admissible, and the signature-change reader
verification is kept verbatim. The legacy ReggeSolver::solve W-space updater
(sign-frozen, 1e-12 floor, 5% trust region) is a separate parametrization on
a path stage 2 does not use — flagged on the PR as its own follow-up.
